Edited backend.
Ensured that it is impossible to create an event in the past for start time and end time.

Added the cancel and re-open button and moved it into update.html, so a user can only cancel or re-open an event if they are in the update event page.

Added cancel order button to set an order status to cancelled from active. Inactive order status has also been implemented, where the live_status() function also updates all orders to be active or inactive.